Antaeus In this commit I am trying to improve the functionality of the project. I have implemented a logic in which the sysadmin can set the range of the database items that the Worker will handle though the use of the API (/rest/config?dbstart=?&dbend=?) or check the current range (/rest/info). I know this might not be the optimal solution, but it will help with the distributed functionality of the project.
AntaeusDal.kt
fetchInvoices: in case of an input the new function will be used in order to select only the rows that are needed.

AntaeusRest.kt
Added info path which returns the range of the database that this worker handles
Added config path so that the user can specify the range that this worker will handle
Improved /rest/v1/invoices/{:id} by adding a try/catch statement to handle bad input cases

BillingService.kt
Added to more error codes (0 success, 1 failed, 2 network error, 3 error customer not found , 4 error currency mismatch) and improved the code so that the user will get a proper and more complete status update

InvoiceService.kt
Added variables to point the database range that the worker will handle

InvoiceStatus.kt
Added two more error codes so that in case of an error that a manual action is needed (Customer not found, Currency mismatch) the database will be updated accordingly.

README.md
First attempt